requirements.txt
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyQt6
qrcode[pil]
Pillow
kivy
opencv-python
requests
